some

+++ Assets/Survive/Scripts/Creatures/CreatureData/Components/MovementComponent.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "MovementComponent", menuName = "Components/Move")]
 public class MovementComponent : ScriptableObject
 {
-    public float WalkSpeed;
+    public int WalkSpeed;
     public AudioClip[] StepSounds;
     public float StepInterval;
 }
